enable SNO and 3-node compact

+++ PAO-config/cleanup.sh
@@ -35,14 +35,24 @@ echo "Next, remove node labels and MCP ${MCP}"
 prompt_continue 
 
 echo "deleting label for $WORKER_LIST ..."
-for worker in $WORKER_LIST; do
-    oc label --overwrite node ${worker} node-role.kubernetes.io/${MCP}-
-done
+if [ "${MCP}" != "master" ]; then
+    # this is STANDARD cluster. Do it.
+    for worker in $WORKER_LIST; do
+        oc label --overwrite node ${worker} node-role.kubernetes.io/${MCP}-
+    done
+fi
 
 ##### Remove MCP ######
-if oc get mcp $MCP 2>/dev/null; then
-    oc delete -f ${MANIFEST_DIR}/mcp-${MCP}.yaml
-    echo "deleted mcp for ${MCP}: done"
+if [ "${MCP}" != "master" ]; then
+    # this is STANDARD cluster. Do it.
+    if oc get mcp $MCP 2>/dev/null; then
+        oc delete -f ${MANIFEST_DIR}/mcp-${MCP}.yaml
+        echo "deleted mcp for ${MCP}: done"
+    fi
+else
+    # this is non-standard cluster that uses mcp master. Just remove the label.
+    oc label --overwrite mcp ${MCP} machineconfiguration.openshift.io/role-
+
 fi
